feat(menuPrice) : 메뉴 가격 VO 추가

+++ src/main/java/kitchenpos/exception/IllegalPriceException.java
@@ -1,11 +1,13 @@
 package kitchenpos.exception;
 
+import java.math.BigDecimal;
+
 public class IllegalPriceException extends IllegalArgumentException {
     public IllegalPriceException(String descriptions, String p) {
         super(descriptions+ " " + p);
     }
 
-    public IllegalPriceException(String descriptions, Long price) {
+    public IllegalPriceException(String descriptions, BigDecimal price) {
         super(descriptions + " " + price);
     }
 }
